docker: improve Docker performance
Currently, the cache is reset _every_ build, which is very ineffective.
This happens because entire directories are copied, so if only a single
file in that directory has been changed, all commands that use any of
those files must reexecute.

Now, single files are added only when required. This improves cache hit
times and vastly improves build performance.